CPUC-2989, CPUC-2990-MASKL and MASKR instrinsics returning wrong values.

+++ test/f08_correct/src/bitint01.f08
@@ -3,15 +3,16 @@
 !
 ! F2008 Compliance Tests: Bit-intrinsics test cases
 !
-! Date of Modification: Tue Feb 11 20:45:41 IST 2020
+! Date of Modification: Thu Mar 19 10:54:19 IST 2020
+!
 ! Purpose: Test operation of BIT_SIZE intrinsic
 !
 PROGRAM BITINT01
   IMPLICIT NONE
   INTEGER, PARAMETER :: N = 1
   LOGICAL EXP(N), RES(N)
-  INTEGER :: I = 123
-  INTEGER :: SIZE
+  INTEGER(KIND=4) :: I = 123
+  INTEGER(KIND=4) :: SIZE
 
   SIZE = BIT_SIZE(I)
   PRINT *, SIZE
